move notifications under settings
