+++ assets/src/routes/settingsRoutesConst.tsx
@@ -1,3 +1,5 @@
+import { Breadcrumb } from '@pluralsh/design-system'
+
 export const SETTINGS_REL_PATH = 'settings' as const
 export const SETTINGS_ABS_PATH = `/${SETTINGS_REL_PATH}` as const
 
@@ -12,6 +14,23 @@ export const GLOBAL_SETTINGS_ABS_PATH =
 export const PROJECT_SETTINGS_REL_PATH = 'projects' as const
 export const PROJECT_SETTINGS_ABS_PATH =
   `${SETTINGS_ABS_PATH}/${PROJECT_SETTINGS_REL_PATH}` as const
+
+export const NOTIFICATIONS_REL_PATH = 'notifications' as const
+export const NOTIFICATIONS_ABS_PATH =
+  `/${SETTINGS_REL_PATH}/${NOTIFICATIONS_REL_PATH}` as const
+export const NOTIFICATIONS_ROUTERS_REL_PATH = 'routers' as const
+export const NOTIFICATIONS_ROUTERS_ABS_PATH =
+  `${NOTIFICATIONS_ABS_PATH}/${NOTIFICATIONS_ROUTERS_REL_PATH}` as const
+export const NOTIFICATIONS_SINKS_REL_PATH = 'sinks' as const
+export const NOTIFICATIONS_SINKS_ABS_PATH =
+  `${NOTIFICATIONS_ABS_PATH}/${NOTIFICATIONS_SINKS_REL_PATH}` as const
+
+export const NOTIFICATIONS_BASE_CRUMBS = [
+  {
+    label: 'Notifications',
+    url: NOTIFICATIONS_ABS_PATH,
+  },
+] as const satisfies readonly Breadcrumb[]
 
 export const AUDITS_REL_PATH = 'audits' as const
 export const AUDITS_ABS_PATH =
